add bert api
